Go back to `oncalls` endpoint
The pdjs library parses the `resource` key based on the _first_ resource
found after the endpoint path. This means that using
`/schedules/{id}/users`, the library will try to set `resource` as
whatever the value is with the key of `schedules`, which the endpoint
does not provide. It actually returns `users` instead.

The `oncalls` endpoint does in fact return a list of `oncalls` so we'll
go back to using that.